Return early if not using Genesis
Because we changed the init hook from a Genesis-specific one, we now
need this guard or the plugin will fatal error if the user activates it
when a Genesis theme is not active.

Before, doing that would have no effect because the instantiate method
was hooked to a Genesis-specific action, and those only exist if
Genesis is active.