Fix term menus to display saved term (#32)
In #26, we stripped $term from term_edit_form to work around a
phpcs warning about unused arguments:

https://github.com/studiopress/genesis-simple-menus/pull/26/files#diff-d27338949948e867d1f6cb42c1f46c0fc668e9dcc90a93789a73d3c42dc46164R109

But the $term argument is actually used in the template
(term-edit-field.php). Stripping it results in menus being saved but
the saved one not being selected as active when the page refreshes.

We can add a phpcs:ignore comment instead so that phpcs is happy and
functionality is not affected.